Modify to the lib/main structure.
