+++ dep_iu.ps1
@@ -1,0 +1,2 @@
+npm i
+cargo install --path ./src-tauri
