+++ dep_iu.sh
@@ -1,0 +1,3 @@
+#!/bin/sh
+npm i
+cargo install --path ./src-tauri
