+++ src-tauri/Cargo.toml
@@ -7,7 +7,10 @@ license = "MIT"
 repository = "https://github.com/VPKSoft/PasswordKeeper"
 default-run = "password_keeper"
 edition = "2021"
-rust-version = "1.59"
+
+[lib]
+name = "app_lib"
+crate-type = ["lib", "cdylib", "staticlib"]
 
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
